pyproject.toml
[project]
name = "ai-meishi"
version = "0.1.0"
description = "AI business card analysis backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9",
  "pydantic>=2.8",
  "pydantic-settings>=2.3",
  "openai>=1.0.0",
  "pypdf>=4.2.0",
  "pillow>=10.0.0",
  "reportlab>=4.2.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "websockets>=12.0",
  "requests>=2.31.0",
  "opencv-python-headless>=4.10.0.84",
  "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "ruff>=0.6",
  "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.services", "app.models"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
